# openapi.yaml
openapi: 3.0.1
info:
  title: Familie Utvidet Barnetrygd
  description: "API for å finne personer/perioder med utvidet barnetrygd innenfor et gitt år.\n\nPåkrevd scope-verdi for bruk av API'et er `nav:familie/v1/barnetrygd/utvidet`."
  version: "1.0"
servers:
  - url: template-value
tags:
  - name: Personer
    description: ""
  - name: Perioder
    description: ""

paths:
  /personer:
    get:
      tags:
        - Personer
      summary: Finn personer med utvidet barnetrygd.
      description: |
        Operasjonen henter en liste over personer som har hatt utvidet barnetrygd i det aktuelle skatteåret.
      operationId: finnPersonerMedUtvidetBarnetrygd
      parameters:
        - name: aar
          in: query
          description: 'Format - 4 siffer. Året det ønskes opplysninger om. Kan være inneværende, tidligere eller fremtidige år.'
          required: true
          schema:
            type: string
        - name: Authorization
          in: header
          description: 'Format - byte. Standard Authorization header som inneholder et Base64-kodet Bearer-token (JWT-aksesstoken) utstedt av Maskinporten (Difi). Verdien må starte med "Bearer ", deretter følger tokenstrengen.'
          required: true
          schema:
            type: string
      responses:
        200:
          description: Returnerer en liste over personer som har hatt utvidet barnetrygd i løpet av det aktuelle året.
          content:
            application/json;charset=UTF-8:
              schema:
                $ref: '#/components/schemas/PersonerResponse'
        '401':
          description: "Unauthorized"
        '403':
          description: "Forbidden"
        '500':
          description: "Internal Server Error"
          content:
            application/json;charset=UTF-8: { }

  /perioder:
    post:
      tags:
        - Perioder
      summary: Hent perioder med utvidet barnetrygd for personer.
      description: |
        Operasjonen henter en liste over perioder med utvidet barnetrygd for hver person innenfor det aktuelle skatteåret.
      operationId: hentPerioderMedUtvidetBarnetrygd
      parameters:
        - name: aar
          in: query
          description: 'Format - 4 siffer. Året det ønskes opplysninger om. Kan være inneværende, tidligere eller fremtidige år.'
          required: true
          schema:
            type: string
        - name: Authorization
          in: header
          description: 'Format - byte. Standard Authorization header som inneholder et Base64-kodet Bearer-token (JWT-aksesstoken) utstedt av Maskinporten (Difi). Verdien må starte med "Bearer ", deretter følger tokenstrengen.'
          required: true
          schema:
            type: string
      requestBody:
        content:
          'application/json':
            schema:
              $ref: '#/components/schemas/PerioderRequest'
        required: true
      responses:
        200:
          description: Returnerer en liste over perioder med utvidet barnetrygd for hver person innenfor det aktuelle året.
          content:
            application/json;charset=UTF-8:
              schema:
                $ref: '#/components/schemas/PerioderResponse'
        '401':
          description: "Unauthorized"
        '403':
          description: "Forbidden"
        '500':
          description: "Internal Server Error"
          content:
            application/json;charset=UTF-8: { }

components:
  schemas:
    PersonerResponse:
      title: PersonerResponse
      type: object
      properties:
        brukere:
          type: array
          items:
            $ref: '#/components/schemas/Person'

    PerioderRequest:
      title: PerioderRequest
      properties:
        identer:
          type: array
          items:
            type: string
          description: Liste over fødselsnumre det ønskes opplysninger om.
      required:
        - identer

    PerioderResponse:
      title: PerioderResponse
      properties:
        brukere:
          type: array
          items:
            $ref: '#/components/schemas/Person'

    Person:
      title: Person
      properties:
        ident:
          type: string
          description: Person identifikator
        sisteVedtakPaaIdent:
          type: string
          description: Tidspunkt for siste vedtak (systemtidspunkt)
          format: date-time
        perioder:
          type: array
          items:
            $ref: '#/components/schemas/Periode'
      required:
        - ident
        - sisteVedtakPaaIdent

    Periode:
      title: Periode
      type: object
      properties:
        fraMaaned:
          type: string
          description: Første måned i perioden.
          format: year-month
        tomMaaned:
          type: string
          description: Den siste måneden i perioden
          format: year-month
        maxDelingsprosent:
          type: string
          enum: [100, 50, usikker]
          description: For perioder som løper i nytt fagsystem, vil "maxDelingsprosent" alltid være "50" eller "100". "usikker" tilsvarer kode "3" i gammelt fagsystem
      description: Representerer en tidsperiode gitt ved en fra-og-med-måned og en
        valgfri til-og-med-måned, og som i tillegg inneholder en opplysning ang. maks delingsprosent i perioden
      required:
        - fraMaaned
        - maxDelingsprosent

  securitySchemes:
    maskinporten:
      type: http
      scheme: bearer
      description: Format - byte. Standard Authorization header som inneholder et
        Base64-kodet Bearer-token (JWT-aksesstoken) utstedt av Maskinporten (Difi).
        Verdien må starte med "Bearer ", deretter følger tokenstrengen.

security:
  - maskinporten: []