# src/main/kotlin/no/nav/familie/ba/skatteetaten/model/PerioderResponse.kt
package org.openapitools.model

import java.util.Objects
import com.fasterxml.jackson.annotation.JsonProperty
import org.openapitools.model.Person
import javax.validation.constraints.DecimalMax
import javax.validation.constraints.DecimalMin
import javax.validation.constraints.Max
import javax.validation.constraints.Min
import javax.validation.constraints.NotNull
import javax.validation.constraints.Pattern
import javax.validation.constraints.Size
import javax.validation.Valid

/**
 * 
 * @param brukere 
 */
data class PerioderResponse(

    @field:Valid
    @field:JsonProperty("brukere") val brukere: kotlin.collections.List<Person>? = null
) {

}